+++ Client/Components/Question.jsx
@@ -1,20 +1,17 @@
-// import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
-// import React, { useEffect, useState } from "react";
+import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
+import React, { useEffect, useState } from "react";
 
-// const Question = (props) => {
-
-//     return(
-//         <form style={{border: '2px solid green'}}className="questionBox">
-//             {/* <h1>{props.currentQuestion}</h1> */}
-//             <input type="radio" id="never" name="answer" value="0" />
-//             <label hmtlFor="never">Never</label><br></br>
-//             <input type="radio" id="sometimes" name="answer" value="2" />
-//             <label htmlFor="sometimes">Sometimes</label><br></br>
-//             <input type="radio" id="always" name="answer" value="4" />
-//             <label htmlFor="always">Always</label><br></br>
-//             {/* <button>NEXT</button> */}
-//         </form>
-//     )
-// }
-
-// export default Question;
+function Question({ question, handleChange, index }) {
+    return (
+        <div>
+            <div>{question.text}</div>
+            <input type="radio" id="never" name={question.name} value={0} index={index} onClick={handleChange} />
+                <label hmtlFor="never">Never</label><br></br>
+            <input type="radio" id="sometimes" name={question.name} value={2} index={index} onClick={handleChange} />
+                <label htmlFor="sometimes">Sometimes</label><br></br>
+            <input type="radio" id="always" name={question.name} value={4} index={index} onClick={handleChange} />
+                <label htmlFor="always">Always</label><br></br>
+        </ div>
+    );
+}
+export default Question;
