+++ Client/Components/Questionnaire.jsx
@@ -1,5 +1,6 @@
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import React, { useEffect, useState } from "react";
+import Question from "./Question.jsx";
 
 function Questionnaire() {
     // currentQuestion is a number that corresponds to an index in the answers array
@@ -84,18 +85,6 @@ function Questionnaire() {
   );
 }
 
-function Question({ question, handleChange, index }) {
-    return (
-        <div>
-            <div>{question.text}</div>
-            <input type="radio" id="never" name={question.name} value={0} index={index} onClick={handleChange} />
-                <label hmtlFor="never">Never</label><br></br>
-            <input type="radio" id="sometimes" name={question.name} value={2} index={index} onClick={handleChange} />
-                <label htmlFor="sometimes">Sometimes</label><br></br>
-            <input type="radio" id="always" name={question.name} value={4} index={index} onClick={handleChange} />
-                <label htmlFor="always">Always</label><br></br>
-        </ div>
-    );
-}
+
 
 export default Questionnaire;
