+++ Client/Containers/App.jsx
@@ -1,14 +1,10 @@
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import React, { useEffect, useState } from "react";
-import Form from "../Components/Form.jsx";
-// import Question from "../Components/Question.jsx";
 import Questionnaire from "../Components/Form.jsx";
 // import '../Sass/styles.css';
 
 
-const App = () => {
-    // const [question, setQuestion] = useState()
-    
+const App = () => {    
     return(
         <div className = "blue">
             <Questionnaire />
